+++ nokhwa-bindings-macos/src/lib.rs
@@ -17,14 +17,40 @@
 #![allow(clippy::not_unsafe_ptr_arg_deref)]
 #[cfg(any(target_os = "macos", target_os = "ios"))]
 mod internal {
-    use std::{borrow::Cow, ffi::{c_float, c_void, CStr}, sync::Arc};
+    use std::{
+        borrow::Cow,
+        collections::BTreeMap,
+        ffi::{c_float, c_void},
+    };
 
-    use av_foundation::{capture_device::{AVCaptureDevice, AVCaptureDeviceType, AVCaptureDeviceFormat, AVCaptureDevicePosition}, capture_input::AVCaptureDeviceInput, capture_session::AVCaptureSession, media_format::{AVMediaType, AVMediaTypeText, AVMediaTypeTimecode, AVMediaTypeVideo}};
-    use block::ConcreteBlock;
+    #[cfg(target_os = "ios")]
+    use av_foundation::capture_device::{
+        AVCaptureDeviceTypeBuiltInDualCamera, AVCaptureDeviceTypeBuiltInTelephotoCamera,
+        AVCaptureDeviceTypeBuiltInTrueDepthCamera, AVCaptureDeviceTypeBuiltInUltraWideCamera,
+    };
+    use av_foundation::{
+        capture_device::{
+            AVCaptureDevice, AVCaptureDeviceFormat, AVCaptureDevicePosition,
+            AVCaptureDevicePositionUnspecified, AVCaptureDeviceType,
+            AVCaptureDeviceTypeBuiltInWideAngleCamera, AVCaptureDeviceTypeContinuityCamera,
+            AVCaptureDeviceTypeDeskViewCamera, AVCaptureDeviceTypeExternalUnknown,
+            AVCaptureExposureDurationCurrent, AVCaptureISOCurrent, AVCaptureWhiteBalanceGains,
+        },
+        capture_device_discovery_session::AVCaptureDeviceDiscoverySession,
+        capture_input::AVCaptureDeviceInput,
+        capture_output_base::AVCaptureOutput,
+        capture_session::{AVCaptureConnection, AVCaptureSession},
+        capture_video_data_output::AVCaptureVideoDataOutputSampleBufferDelegate,
+        media_format::{AVMediaType, AVMediaTypeText, AVMediaTypeTimecode, AVMediaTypeVideo},
+    };
+    use core_foundation::base::TCFType;
     use core_media::{
-        format_description::{CMFormatDescriptionGetMediaSubType, CMFormatDescriptionRef, CMVideoFormatDescriptionGetDimensions},
+        format_description::{
+            CMFormatDescriptionGetMediaSubType, CMFormatDescriptionRef,
+            CMVideoFormatDescriptionGetDimensions,
+        },
         sample_buffer::{
-            CMSampleBufferGetFormatDescription, CMSampleBufferGetImageBuffer, CMSampleBufferRef,
+            CMSampleBufferGetFormatDescription, CMSampleBufferGetImageBuffer, CMSampleBuffer, CMSampleBufferRef,
         },
         time::CMTime,
         OSType,
@@ -33,23 +59,35 @@ mod internal {
         image_buffer::CVImageBufferRef,
         pixel_buffer::{
             CVPixelBufferGetBaseAddress, CVPixelBufferGetDataSize, CVPixelBufferLockBaseAddress,
-            CVPixelBufferUnlockBaseAddress,
+            CVPixelBufferUnlockBaseAddress, CVPixelBuffer,
         },
     };
     use dispatch::{Queue, QueueAttribute};
-    use flume::{Sender, Receiver};
+    use flume::{Receiver, Sender};
     use four_cc::FourCC;
     use nokhwa_core::{
         error::NokhwaError,
         pixel_format::GRAY,
         types::{
-            ApiBackend, CameraControl, CameraFormat, CameraIndex, CameraInfo, ControlValueDescription, ControlValueSetter, KnownCameraControl, KnownCameraControlFlag, Resolution
+            ApiBackend, CameraControl, CameraFormat, CameraIndex, CameraInfo,
+            ControlValueDescription, ControlValueSetter, KnownCameraControl,
+            KnownCameraControlFlag, Resolution,
         },
     };
     use objc2::{
-        class, declare::ClassDecl, declare_class, ffi::{Nil, BOOL, NO, YES}, msg_send, runtime::{AnyObject, Class, Protocol, Sel}, sel
+        class,
+        declare::ClassDecl,
+        DeclaredClass,
+        declare_class, extern_methods,
+        ffi::{Nil, BOOL, NO, YES},
+        msg_send, msg_send_id, mutability,
+        rc::{Allocated, Id},
+        runtime::{AnyObject, Class, Protocol, Sel},
+        sel, ClassType,
     };
-    use objc2_foundation::{ns_string, CGFloat, CGPoint, NSArray, NSInteger, NSObject, NSString};
+    use objc2_foundation::{
+        ns_string, CGFloat, CGPoint, NSArray, NSInteger, NSObject, NSObjectProtocol, NSString,
+    };
     use once_cell::sync::Lazy;
 
     #[allow(non_upper_case_globals)]
@@ -60,152 +98,152 @@ mod internal {
     pub type CompressionData<'a> = (Cow<'a, [u8]>, FourCC);
     pub type DataPipe<'a> = (Sender<CompressionData<'a>>, Receiver<CompressionData<'a>>);
 
-    static CALLBACK_CLASS: Lazy<&'static Class> = Lazy::new(|| {
-        {
-            let mut decl = ClassDecl::new("MyCaptureCallback", class!(NSObject)).unwrap();
+    pub struct DelegateIvars {}
 
-            decl.add_ivar::<*const c_void>("_arcmutptr"); 
+    declare_class!(
+        struct Delegate;
 
-            extern "C" fn my_callback_get_arcmutptr(this: &AnyObject, _: Sel) -> *const c_void {
-                let ivar = CALLBACK_CLASS::get("MyCaptureCallback").unwrap().instance_variable("_arcmutptr").unwrap();
-                unsafe { ivar.load::<*const c_void>(this) }.get()
-            }
-            extern "C" fn my_callback_set_arcmutptr(
-                this: &mut AnyObject,
-                _: Sel,
-                new_arcmutptr: *const c_void,
+        unsafe impl ClassType for Delegate {
+            type Super = NSObject;
+            type Mutability = mutability::Mutable;
+            const NAME: &'static str = "OutputSampleBufferDelegate";
+        }
+
+        impl DeclaredClass for Delegate {
+            type Ivars = DelegateIvars;
+        }
+
+        unsafe impl NSObjectProtocol for Delegate {}
+
+        unsafe impl AVCaptureVideoDataOutputSampleBufferDelegate for Delegate {
+            #[method(captureOutput:didOutputSampleBuffer:fromConnection:)]
+            unsafe fn capture_output_did_output_sample_buffer(
+                &self,
+                _capture_output: &AVCaptureOutput,
+                sample_buffer: CMSampleBufferRef,
+                _connection: &AVCaptureConnection,
             ) {
-                let ivar = CALLBACK_CLASS::get("MyCaptureCallback").unwrap().instance_variable("_arcmutptr").unwrap();
-                unsafe { ivar.load::<*const c_void>(this) }.set(new_arcmutptr);
-            }
-
-            // Delegate compliance method
-            // SAFETY: This assumes that the buffer byte size is a u8. Any other size will cause unsafety.
-            #[allow(non_snake_case)]
-            #[allow(non_upper_case_globals)]
-            extern "C" fn capture_out_callback(
-                this: &mut AnyObject,
-                _: Sel,
-                _: *mut AnyObject,
-                didOutputSampleBuffer: CMSampleBufferRef,
-                _: *mut AnyObject,
-            ) {
-                let format = unsafe { CMSampleBufferGetFormatDescription(didOutputSampleBuffer) };
-                let media_subtype = unsafe { CMFormatDescriptionGetMediaSubType(format) };
-                let media_subtype_fcc = FourCC::from(media_subtype);
-                println!("media_subtype_fcc: {:?}", media_subtype_fcc);
-
-                let image_buffer: CVImageBufferRef =
-                    unsafe { CMSampleBufferGetImageBuffer(didOutputSampleBuffer) };
-                unsafe {
-                    CVPixelBufferLockBaseAddress(image_buffer, 0);
-                };
-
-                let buffer_length = unsafe { CVPixelBufferGetDataSize(image_buffer) };
-                let buffer_ptr = unsafe { CVPixelBufferGetBaseAddress(image_buffer) };
-                let buffer_as_vec = unsafe {
-                    std::slice::from_raw_parts_mut(buffer_ptr as *mut u8, buffer_length as usize)
-                        .to_vec()
-                };
-
-                unsafe { CVPixelBufferUnlockBaseAddress(image_buffer, 0) };
-                // oooooh scarey unsafe
-                // AAAAAAAAAAAAAAAAAAAAAAAAA
-                // https://c.tenor.com/0e_zWtFLOzQAAAAC/needy-streamer-overload-needy-girl-overdose.gif
-                let bufferlck_cv: *const c_void = unsafe { msg_send![this, bufferPtr] };
-                let buffer_sndr: Arc<Sender<(Vec<u8>, FourCC)>> = unsafe {
-                    let ptr = bufferlck_cv.cast::<Sender<(Vec<u8>, FourCC)>>();
-                    Arc::from_raw(ptr)
-                };
-                if let Err(_) = buffer_sndr.send((buffer_as_vec, GRAY)) {
-                    return;
+                let sample_buffer = CMSampleBuffer::wrap_under_get_rule(sample_buffer);
+                if let Some(image_buffer) = sample_buffer.get_image_buffer() {
+                    if let Some(pixel_buffer) = image_buffer.downcast::<CVPixelBuffer>() {
+                        println!("pixel buffer: {:?}", pixel_buffer);
+                    }
                 }
-                std::mem::forget(buffer_sndr);
             }
 
-            #[allow(non_snake_case)]
-            extern "C" fn capture_drop_callback(
-                _: &mut AnyObject,
-                _: Sel,
-                _: *mut AnyObject,
-                _: *mut AnyObject,
-                _: *mut AnyObject,
-            ) {
-            }
-
-            unsafe {
-                decl.add_method(
-                    sel!(bufferPtr),
-                    my_callback_get_arcmutptr as extern "C" fn(&AnyObject, Sel) -> *const c_void,
-                );
-                decl.add_method(
-                    sel!(SetBufferPtr:),
-                    my_callback_set_arcmutptr as extern "C" fn(&mut AnyObject, Sel, *const c_void),
-                );
-                decl.add_method(
-                    sel!(captureOutput:didOutputSampleBuffer:fromConnection:),
-                    capture_out_callback
-                        as extern "C" fn(
-                            &mut AnyObject,
-                            Sel,
-                            *mut AnyObject,
-                            CMSampleBufferRef,
-                            *mut AnyObject,
-                        ),
-                );
-                decl.add_method(
-                    sel!(captureOutput:didDropSampleBuffer:fromConnection:),
-                    capture_drop_callback
-                        as extern "C" fn(
-                            &mut AnyObject,
-                            Sel,
-                            *mut AnyObject,
-                            *mut AnyObject,
-                            *mut AnyObject,
-                        ),
-                );
-
-                decl.add_protocol(
-                    Protocol::get("AVCaptureVideoDataOutputSampleBufferDelegate").unwrap(),
-                );
-            }
-
-            decl.register()
+            #[method(captureOutput:didDropSampleBuffer:fromConnection:)]
+            unsafe fn capture_output_did_drop_sample_buffer(
+            &self,
+            _capture_output: &AVCaptureOutput,
+            sample_buffer: CMSampleBufferRef,
+            _connection: &AVCaptureConnection,
+            ) {}
         }
-    });
 
-    pub fn request_permission_with_callback(callback: impl Fn(bool) + Send + Sync + 'static) {
-        let cls = class!(AVCaptureDevice);
-
-        let wrapper = move |bool: BOOL| {
-            callback(bool == YES);
-        };
-
-        let objc_fn_block: ConcreteBlock<(BOOL,), (), _> = ConcreteBlock::new(wrapper);
-        let objc_fn_pass = objc_fn_block.copy();
-
-        unsafe {
-            let _: () = msg_send![cls, requestAccessForMediaType:(AVMediaTypeVideo.clone()) completionHandler:objc_fn_pass];
+        unsafe impl Delegate {
+            #[method_id(init)]
+            fn init(this: Allocated<Self>) -> Option<Id<Self>> {
+                let this = this.set_ivars(DelegateIvars {});
+                unsafe { msg_send_id![super(this), init] }
+            }
         }
-    }
+    );
 
-    pub fn current_authorization_status() -> AVAuthorizationStatus {
-        let cls = class!(AVCaptureDevice);
-        let status: AVAuthorizationStatus = unsafe {
-            msg_send![cls, authorizationStatusForMediaType:AVMediaType::Video.into_ns_str()]
-        };
-        status
-    }
+    extern_methods!(
+        unsafe impl Delegate {
+            #[method_id(new)]
+            pub fn new() -> Id<Self>;
+        }
+    );
+
+    // Delegate compliance method
+    // SAFETY: This assumes that the buffer byte size is a u8. Any other size will cause unsafety.
+    // #[allow(non_snake_case)]
+    // #[allow(non_upper_case_globals)]
+    // extern "C" fn capture_out_callback(
+    //     this: &mut AnyObject,
+    //     _: Sel,
+    //     _: *mut AnyObject,
+    //     didOutputSampleBuffer: CMSampleBufferRef,
+    //     _: *mut AnyObject,
+    // ) {
+    //     let format = unsafe { CMSampleBufferGetFormatDescription(didOutputSampleBuffer) };
+    //     let media_subtype = unsafe { CMFormatDescriptionGetMediaSubType(format) };
+    //     let media_subtype_fcc = FourCC::from(media_subtype);
+    //     println!("media_subtype_fcc: {:?}", media_subtype_fcc);
+
+    //     let image_buffer: CVImageBufferRef =
+    //         unsafe { CMSampleBufferGetImageBuffer(didOutputSampleBuffer) };
+    //     unsafe {
+    //         CVPixelBufferLockBaseAddress(image_buffer, 0);
+    //     };
+
+    //     let buffer_length = unsafe { CVPixelBufferGetDataSize(image_buffer) };
+    //     let buffer_ptr = unsafe { CVPixelBufferGetBaseAddress(image_buffer) };
+    //     let buffer_as_vec = unsafe {
+    //         std::slice::from_raw_parts_mut(buffer_ptr as *mut u8, buffer_length as usize)
+    //             .to_vec()
+    //     };
+
+    //     unsafe { CVPixelBufferUnlockBaseAddress(image_buffer, 0) };
+    //     // oooooh scarey unsafe
+    //     // AAAAAAAAAAAAAAAAAAAAAAAAA
+    //     // https://c.tenor.com/0e_zWtFLOzQAAAAC/needy-streamer-overload-needy-girl-overdose.gif
+    //     let bufferlck_cv: *const c_void = unsafe { msg_send![this, bufferPtr] };
+    //     let buffer_sndr: Arc<Sender<(Vec<u8>, FourCC)>> = unsafe {
+    //         let ptr = bufferlck_cv.cast::<Sender<(Vec<u8>, FourCC)>>();
+    //         Arc::from_raw(ptr)
+    //     };
+    //     if let Err(_) = buffer_sndr.send((buffer_as_vec, GRAY)) {
+    //         return;
+    //     }
+    //     std::mem::forget(buffer_sndr);
+    // }
 
     pub fn query_avfoundation() -> Result<Vec<CameraInfo>, NokhwaError> {
-        Ok(AVCaptureDeviceDiscoverySession::new(vec![
-            AVCaptureDevice::UltraWide,
-            AVCaptureDevice::WideAngle,
-            AVCaptureDevice::Telephoto,
-            AVCaptureDevice::TrueDepth,
-            AVCaptureDevice::ExternalUnknown,
-        ])?
-        .devices())
+        #[cfg(any(target_os = "macos"))]
+        let device_types: Vec<&AVCaptureDeviceType> = vec![
+            AVCaptureDeviceTypeBuiltInWideAngleCamera,
+            AVCaptureDeviceTypeContinuityCamera,
+            AVCaptureDeviceTypeDeskViewCamera,
+            AVCaptureDeviceTypeExternalUnknown,
+        ];
+
+        #[cfg(any(target_os = "ios"))]
+        let device_types: Vec<&AVCaptureDeviceType> = vec![
+            AVCaptureDeviceTypeBuiltInUltraWideCamera,
+            AVCaptureDeviceTypeBuiltInWideAngleCamera,
+            AVCaptureDeviceTypeBuiltInTelephotoCamera,
+            AVCaptureDeviceTypeBuiltInDualCamera,
+            AVCaptureDeviceTypeBuiltInTrueDepthCamera,
+            AVCaptureDeviceTypeExternalUnknown,
+        ];
+        let device_types_nsarray = NSArray::new();
+        device_types.iter().for_each(|device_type| unsafe {
+            device_types_nsarray = device_types_nsarray.arrayByAddingObject(*device_type);
+        });
+        let discovery_session =
+            AVCaptureDeviceDiscoverySession::discovery_session_with_device_types(
+                &device_types_nsarray,
+                AVMediaTypeVideo.clone(),
+                AVCaptureDevicePositionUnspecified,
+            );
+        let devices = discovery_session.devices();
+        let cameras = devices.into_iter().map(|device| {
+            CameraInfo(
+                device.localized_name().to_string(),
+                device.unique_id().to_string(),
+                device.manufacturer().to_string(),
+                format!(
+                    "{}: {} - {}, {:?}",
+                    device.manufacturer(),
+                    device.model_id(),
+                    device.device_type(),
+                    device.position()
+                ),
+            )
+        }).collect();
+        Ok(cameras)
     }
 
     pub fn get_raw_device_info(index: CameraIndex, device: AVCaptureDevice) -> CameraInfo {
@@ -223,98 +261,8 @@ mod internal {
         CameraInfo::new(name.as_ref(), &description, &uuid.clone(), index)
     }
 
-    #[derive(Copy, Clone, Debug, Hash, Ord, PartialOrd, Eq, PartialEq)]
-    #[repr(isize)]
-    pub enum AVAuthorizationStatus {
-        NotDetermined = 0,
-        Restricted = 1,
-        Denied = 2,
-        Authorized = 3,
-    }
-
-    pub struct AVCaptureVideoCallback {
-        delegate: *mut AnyObject,
-        queue: Queue,
-    }
-
-    impl AVCaptureVideoCallback {
-        pub fn new(
-            device_spec: &CStr,
-            buffer: &Arc<Sender<(Vec<u8>, FourCC)>>,
-        ) -> Result<Self, NokhwaError> {
-            let cls = &CALLBACK_CLASS as &Class;
-            let delegate: *mut AnyObject = unsafe { msg_send![cls, alloc] };
-            let delegate: *mut AnyObject = unsafe { msg_send![delegate, init] };
-            let buffer_as_ptr = {
-                let arc_raw = Arc::as_ptr(buffer);
-                arc_raw.cast::<c_void>()
-            };
-            unsafe {
-                let _: () = msg_send![delegate, SetBufferPtr: buffer_as_ptr];
-            }
-
-            let queue = unsafe {
-                Queue::create(device_spec.to_str(), QueueAttribute::Serial)
-            };
-
-            Ok(AVCaptureVideoCallback { delegate, queue })
-        }
-
-        pub fn data_len(&self) -> usize {
-            unsafe { msg_send![self.delegate, dataLength] }
-        }
-
-        pub fn inner(&self) -> *mut AnyObject {
-            self.delegate
-        }
-
-        pub fn queue(&self) -> &Queue {
-            &self.queue
-        }
-    }
-
-    impl AVCaptureDeviceDiscoverySession {
-        pub fn new(device_types: Vec<AVCaptureDeviceType>) -> Result<Self, NokhwaError> {
-            let device_types = NSArray::from(device_types);
-            let position = 0 as NSInteger;
-
-            let discovery_session_cls = class!(AVCaptureDeviceDiscoverySession);
-            let discovery_session: *mut AnyObject = unsafe {
-                msg_send![discovery_session_cls, discoverySessionWithDeviceTypes:device_types mediaType:AVMediaTypeVideo position:position]
-            };
-
-            Ok(AVCaptureDeviceDiscoverySession {
-                inner: discovery_session,
-            })
-        }
-
-        pub fn default() -> Result<Self, NokhwaError> {
-            AVCaptureDeviceDiscoverySession::new(vec![
-                AVCaptureDevice::UltraWide,
-                AVCaptureDevice::Telephoto,
-                AVCaptureDevice::ExternalUnknown,
-                AVCaptureDevice::Dual,
-                AVCaptureDevice::DualWide,
-                AVCaptureDevice::Triple,
-            ])
-        }
-
-        pub fn devices(&self) -> Vec<CameraInfo> {
-            let raw_devices = unsafe { msg_send![self.inner, devices] }.to_vec();
-            let mut devices = Vec::with_capacity(raw_devices.length());
-            for (index, device) in raw_devices.iter().enumerate() {
-                devices.push(get_raw_device_info(
-                    CameraIndex::Index(index as u32),
-                    device,
-                ));
-            }
-
-            devices
-        }
-    }
-
     pub struct AVCaptureDeviceWrapper {
-        inner: *mut AVCaptureDevice,
+        inner: AVCaptureDevice,
         device: CameraInfo,
         locked: bool,
     }
@@ -341,7 +289,7 @@ mod internal {
         }
 
         pub fn from_id(id: &str, index_hint: Option<CameraIndex>) -> Result<Self, NokhwaError> {
-            let nsstr_id = NSString::from_str(&id.to_string());
+            let nsstr_id = NSString::from_str(&id.to_string()).as_ref();
             let avfoundation_capture_cls = class!(AVCaptureDevice);
             let capture: *mut AVCaptureDevice =
                 unsafe { msg_send![avfoundation_capture_cls, deviceWithUniqueID: nsstr_id] };
@@ -367,10 +315,10 @@ mod internal {
             &self.device
         }
 
-        pub fn supported_formats_raw(&self) -> Result<Vec<AVCaptureDeviceFormat>, NokhwaError> {
+        pub fn supported_formats_raw(&self) -> Result<Vec<&AVCaptureDeviceFormat>, NokhwaError> {
             unsafe {
-                return msg_send![self.inner, formats]
-                    .to_vec<AVCaptureDeviceFormat>()
+                let raw_formats = self.inner.formats().to_vec().clone();
+                return Ok(raw_formats);
             }
         }
 
@@ -380,14 +328,18 @@ mod internal {
                 .into_iter()
                 .flat_map(|av_fmt| {
                     let resolution = av_fmt.format_description();
-                    av_fmt.video_supported_frame_rate_ranges().iter().map(move |fps_f64| {
-                        let fps = *fps_f64 as isize;
-                        let format = FourCC::from(av_fmt.format_description().get_media_subtype());
+                    av_fmt
+                        .video_supported_frame_rate_ranges()
+                        .iter()
+                        .map(move |fps_f64| {
+                            let fps = *fps_f64 as isize;
+                            let format =
+                                FourCC::from(av_fmt.format_description().get_media_subtype());
 
-                        let resolution =
-                            Resolution::new(resolution.width as u32, resolution.height as u32);
-                        CameraFormat::new(resolution, av_fmt.fourcc, fps)
-                    })
+                            let resolution =
+                                Resolution::new(resolution.width as u32, resolution.height as u32);
+                            CameraFormat::new(resolution, av_fmt.fourcc, fps)
+                        })
                 })
                 .filter(|x| x.frame_rate() != 0)
                 .collect())
@@ -404,7 +356,7 @@ mod internal {
                 });
             }
             let err_ptr: *mut c_void = std::ptr::null_mut();
-            let accepted: BOOL = unsafe { msg_send![self.inner, lockForConfiguration: err_ptr] };
+            let accepted: BOOL = unsafe { msg_send![&self.inner, lockForConfiguration: err_ptr] };
             if !err_ptr.is_null() {
                 return Err(NokhwaError::SetPropertyError {
                     property: "lockForConfiguration".to_string(),
@@ -426,14 +378,14 @@ mod internal {
         pub fn unlock(&mut self) {
             if self.locked {
                 self.locked = false;
-                unsafe { msg_send![self.inner, unlockForConfiguration] }
+                unsafe { msg_send![&self.inner, unlockForConfiguration] }
             }
         }
 
         // thank you ffmpeg
         pub fn set_all(&mut self, descriptor: CameraFormat) -> Result<(), NokhwaError> {
             self.lock()?;
-            let format_list_raw = (unsafe { msg_send![self.inner, formats] })?;
+            let format_list_raw = (unsafe { msg_send![&self.inner, formats] })?;
             let format_list = format_list_raw.to_vec();
             let format_description_sel = sel!(formatDescription);
 
@@ -476,14 +428,14 @@ mod internal {
             let active_video_min_frame_duration = ns_string!("activeVideoMinFrameDuration");
             let active_video_max_frame_duration = ns_string!("activeVideoMaxFrameDuration");
             let _: () =
-                unsafe { msg_send![self.inner, setValue:selected_format forKey:activefmtkey] };
+                unsafe { msg_send![&self.inner, setValue:selected_format forKey:activefmtkey] };
             let min_frame_duration: *mut AnyObject =
                 unsafe { msg_send![selected_range, valueForKey: min_frame_duration] };
             let _: () = unsafe {
-                msg_send![self.inner, setValue:min_frame_duration forKey:active_video_min_frame_duration]
+                msg_send![&self.inner, setValue:min_frame_duration forKey:active_video_min_frame_duration]
             };
             let _: () = unsafe {
-                msg_send![self.inner, setValue:min_frame_duration forKey:active_video_max_frame_duration]
+                msg_send![&self.inner, setValue:min_frame_duration forKey:active_video_max_frame_duration]
             };
             self.unlock();
             Ok(())
@@ -497,18 +449,18 @@ mod internal {
         // 5 => Exposure ISO
         // 6 => Exposure Duration
         pub fn get_controls(&self) -> Result<Vec<CameraControl>, NokhwaError> {
-            let active_format: *mut AnyObject = unsafe { msg_send![self.inner, activeFormat] };
+            let active_format: *mut AnyObject = unsafe { msg_send![&self.inner, activeFormat] };
 
             let mut controls = vec![];
             // get focus modes
 
-            let focus_current: NSInteger = unsafe { msg_send![self.inner, focusMode] };
+            let focus_current: NSInteger = unsafe { msg_send![&self.inner, focusMode] };
             let focus_locked: BOOL =
-                unsafe { msg_send![self.inner, isFocusModeSupported:NSInteger::from(0)] };
+                unsafe { msg_send![&self.inner, isFocusModeSupported:NSInteger::from(0)] };
             let focus_auto: BOOL =
-                unsafe { msg_send![self.inner, isFocusModeSupported:NSInteger::from(1)] };
+                unsafe { msg_send![&self.inner, isFocusModeSupported:NSInteger::from(1)] };
             let focus_continuous: BOOL =
-                unsafe { msg_send![self.inner, isFocusModeSupported:NSInteger::from(2)] };
+                unsafe { msg_send![&self.inner, isFocusModeSupported:NSInteger::from(2)] };
 
             {
                 let mut supported_focus_values = vec![];
@@ -537,8 +489,8 @@ mod internal {
             }
 
             let focus_poi_supported: BOOL =
-                unsafe { msg_send![self.inner, isFocusPointOfInterestSupported] };
-            let focus_poi: CGPoint = unsafe { msg_send![self.inner, focusPointOfInterest] };
+                unsafe { msg_send![&self.inner, isFocusPointOfInterestSupported] };
+            let focus_poi: CGPoint = unsafe { msg_send![&self.inner, focusPointOfInterest] };
 
             controls.push(CameraControl::new(
                 KnownCameraControl::Other(0),
@@ -559,8 +511,8 @@ mod internal {
             ));
 
             let focus_manual: BOOL =
-                unsafe { msg_send![self.inner, isLockingFocusWithCustomLensPositionSupported] };
-            let focus_lenspos: f32 = unsafe { msg_send![self.inner, lensPosition] };
+                unsafe { msg_send![&self.inner, isLockingFocusWithCustomLensPositionSupported] };
+            let focus_lenspos: f32 = unsafe { msg_send![&self.inner, lensPosition] };
 
             controls.push(CameraControl::new(
                 KnownCameraControl::Other(1),
@@ -584,15 +536,15 @@ mod internal {
             ));
 
             // get exposures
-            let exposure_current: NSInteger = unsafe { msg_send![self.inner, exposureMode] };
+            let exposure_current: NSInteger = unsafe { msg_send![&self.inner, exposureMode] };
             let exposure_locked: BOOL =
-                unsafe { msg_send![self.inner, isExposureModeSupported:NSInteger::from(0)] };
+                unsafe { msg_send![&self.inner, isExposureModeSupported:NSInteger::from(0)] };
             let exposure_auto: BOOL =
-                unsafe { msg_send![self.inner, isExposureModeSupported:NSInteger::from(1)] };
+                unsafe { msg_send![&self.inner, isExposureModeSupported:NSInteger::from(1)] };
             let exposure_continuous: BOOL =
-                unsafe { msg_send![self.inner, isExposureModeSupported:NSInteger::from(2)] };
+                unsafe { msg_send![&self.inner, isExposureModeSupported:NSInteger::from(2)] };
             let exposure_custom: BOOL =
-                unsafe { msg_send![self.inner, isExposureModeSupported:NSInteger::from(3)] };
+                unsafe { msg_send![&self.inner, isExposureModeSupported:NSInteger::from(3)] };
 
             {
                 let mut supported_exposure_values = vec![];
@@ -624,8 +576,8 @@ mod internal {
             }
 
             let exposure_poi_supported: BOOL =
-                unsafe { msg_send![self.inner, isExposurePointOfInterestSupported] };
-            let exposure_poi: CGPoint = unsafe { msg_send![self.inner, exposurePointOfInterest] };
+                unsafe { msg_send![&self.inner, isExposurePointOfInterestSupported] };
+            let exposure_poi: CGPoint = unsafe { msg_send![&self.inner, exposurePointOfInterest] };
 
             controls.push(CameraControl::new(
                 KnownCameraControl::Other(2),
@@ -646,7 +598,7 @@ mod internal {
             ));
 
             let expposure_face_driven_supported: BOOL =
-                unsafe { msg_send![self.inner, isFaceDrivenAutoExposureEnabled] };
+                unsafe { msg_send![&self.inner, isFaceDrivenAutoExposureEnabled] };
             let exposure_face_driven: BOOL = unsafe {
                 msg_send![
                     self.inner,
@@ -672,9 +624,9 @@ mod internal {
                 exposure_poi_supported == YES,
             ));
 
-            let exposure_bias: f32 = unsafe { msg_send![self.inner, exposureTargetBias] };
-            let exposure_bias_min: f32 = unsafe { msg_send![self.inner, minExposureTargetBias] };
-            let exposure_bias_max: f32 = unsafe { msg_send![self.inner, maxExposureTargetBias] };
+            let exposure_bias: f32 = unsafe { msg_send![&self.inner, exposureTargetBias] };
+            let exposure_bias_min: f32 = unsafe { msg_send![&self.inner, minExposureTargetBias] };
+            let exposure_bias_max: f32 = unsafe { msg_send![&self.inner, maxExposureTargetBias] };
 
             controls.push(CameraControl::new(
                 KnownCameraControl::Other(4),
@@ -690,7 +642,7 @@ mod internal {
                 true,
             ));
 
-            let exposure_duration: CMTime = unsafe { msg_send![self.inner, exposureDuration] };
+            let exposure_duration: CMTime = unsafe { msg_send![&self.inner, exposureDuration] };
             let exposure_duration_min: CMTime =
                 unsafe { msg_send![active_format, minExposureDuration] };
             let exposure_duration_max: CMTime =
@@ -704,7 +656,7 @@ mod internal {
                     max: exposure_duration_max.value as isize,
                     value: exposure_duration.value as isize,
                     step: 1,
-                    default: unsafe { AVCaptureExposureDurationCurrent.value },
+                    default: unsafe { AVCaptureExposureDurationCurrent.value } as isize,
                 },
                 if exposure_custom == YES {
                     vec![
@@ -717,7 +669,7 @@ mod internal {
                 exposure_custom == YES,
             ));
 
-            let exposure_iso: f32 = unsafe { msg_send![self.inner, ISO] };
+            let exposure_iso: f32 = unsafe { msg_send![&self.inner, ISO] };
             let exposure_iso_min: f32 = unsafe { msg_send![active_format, minISO] };
             let exposure_iso_max: f32 = unsafe { msg_send![active_format, maxISO] };
 
@@ -729,7 +681,7 @@ mod internal {
                     max: exposure_iso_max as f64,
                     value: exposure_iso as f64,
                     step: f32::MIN_POSITIVE as f64,
-                    default: unsafe { AVCaptureISO } as f64,
+                    default: *unsafe { AVCaptureISOCurrent } as f64,
                 },
                 if exposure_custom == YES {
                     vec![
@@ -743,14 +695,13 @@ mod internal {
             ));
 
             // get whiteblaance
-            let white_balance_: NSInteger =
-                unsafe { msg_send![self.inner, whiteBalanceMode] };
+            let white_balance_mode: NSInteger = unsafe { msg_send![&self.inner, whiteBalanceMode] };
             let white_balance_manual: BOOL =
-                unsafe { msg_send![self.inner, isWhiteBalanceModeSupported:NSInteger::from(0)] };
+                unsafe { msg_send![&self.inner, isWhiteBalanceModeSupported:NSInteger::from(0)] };
             let white_balance_auto: BOOL =
-                unsafe { msg_send![self.inner, isWhiteBalanceModeSupported:NSInteger::from(1)] };
+                unsafe { msg_send![&self.inner, isWhiteBalanceModeSupported:NSInteger::from(1)] };
             let white_balance_continuous: BOOL =
-                unsafe { msg_send![self.inner, isWhiteBalanceModeSupported:NSInteger::from(2)] };
+                unsafe { msg_send![&self.inner, isWhiteBalanceModeSupported:NSInteger::from(2)] };
 
             {
                 let mut possible = vec![];
@@ -769,7 +720,7 @@ mod internal {
                     KnownCameraControl::WhiteBalance,
                     "WhiteBalanceMode".to_string(),
                     ControlValueDescription::Enum {
-                        value: white_balance_,
+                        value: white_balance_mode,
                         possible,
                         default: 0,
                     },
@@ -779,11 +730,11 @@ mod internal {
             }
 
             let white_balance_gains: AVCaptureWhiteBalanceGains =
-                unsafe { msg_send![self.inner, deviceWhiteBalanceGains] };
+                unsafe { msg_send![&self.inner, deviceWhiteBalanceGains] };
             let white_balance_default: AVCaptureWhiteBalanceGains =
-                unsafe { msg_send![self.inner, grayWorldDeviceWhiteBalanceGains] };
+                unsafe { msg_send![&self.inner, grayWorldDeviceWhiteBalanceGains] };
             let white_balance_max: AVCaptureWhiteBalanceGains =
-                unsafe { msg_send![self.inner, maxWhiteBalanceGain] };
+                unsafe { msg_send![&self.inner, maxWhiteBalanceGain] };
             let white_balance_gain_supported: BOOL = unsafe {
                 msg_send![
                     self.inner,
@@ -823,14 +774,14 @@ mod internal {
             ));
 
             // get flash
-            let has_torch: BOOL = unsafe { msg_send![self.inner, isTorchAvailable] };
-            let torch_active: BOOL = unsafe { msg_send![self.inner, isTorchActive] };
+            let has_torch: BOOL = unsafe { msg_send![&self.inner, isTorchAvailable] };
+            let torch_active: BOOL = unsafe { msg_send![&self.inner, isTorchActive] };
             let torch_off: BOOL =
-                unsafe { msg_send![self.inner, isTorchModeSupported:NSInteger::from(0)] };
+                unsafe { msg_send![&self.inner, isTorchModeSupported:NSInteger::from(0)] };
             let torch_on: BOOL =
-                unsafe { msg_send![self.inner, isTorchModeSupported:NSInteger::from(1)] };
+                unsafe { msg_send![&self.inner, isTorchModeSupported:NSInteger::from(1)] };
             let torch_auto: BOOL =
-                unsafe { msg_send![self.inner, isTorchModeSupported:NSInteger::from(2)] };
+                unsafe { msg_send![&self.inner, isTorchModeSupported:NSInteger::from(2)] };
 
             {
                 let mut possible = vec![];
@@ -866,8 +817,8 @@ mod internal {
             }
 
             // get low light boost
-            let has_llb: BOOL = unsafe { msg_send![self.inner, isLowLightBoostSupported] };
-            let llb_enabled: BOOL = unsafe { msg_send![self.inner, isLowLightBoostEnabled] };
+            let has_llb: BOOL = unsafe { msg_send![&self.inner, isLowLightBoostSupported] };
+            let llb_enabled: BOOL = unsafe { msg_send![&self.inner, isLowLightBoostEnabled] };
 
             {
                 controls.push(CameraControl::new(
@@ -890,9 +841,9 @@ mod internal {
             }
 
             // get zoom factor
-            let zoom_: CGFloat = unsafe { msg_send![self.inner, videoZoomFactor] };
-            let zoom_min: CGFloat = unsafe { msg_send![self.inner, minAvailableVideoZoomFactor] };
-            let zoom_max: CGFloat = unsafe { msg_send![self.inner, maxAvailableVideoZoomFactor] };
+            let zoom_current: CGFloat = unsafe { msg_send![&self.inner, videoZoomFactor] };
+            let zoom_min: CGFloat = unsafe { msg_send![&self.inner, minAvailableVideoZoomFactor] };
+            let zoom_max: CGFloat = unsafe { msg_send![&self.inner, maxAvailableVideoZoomFactor] };
 
             controls.push(CameraControl::new(
                 KnownCameraControl::Zoom,
@@ -910,9 +861,9 @@ mod internal {
 
             // zoom distortion correction
             let distortion_correction_supported: BOOL =
-                unsafe { msg_send![self.inner, isGeometricDistortionCorrectionSupported] };
+                unsafe { msg_send![&self.inner, isGeometricDistortionCorrectionSupported] };
             let distortion_correction_current_value: BOOL =
-                unsafe { msg_send![self.inner, isGeometricDistortionCorrectionEnabled] };
+                unsafe { msg_send![&self.inner, isGeometricDistortionCorrectionEnabled] };
 
             controls.push(CameraControl::new(
                 KnownCameraControl::Other(6),
@@ -988,7 +939,7 @@ mod internal {
                     }
 
                     let _: () = unsafe {
-                        msg_send![self.inner, setExposureModeCustomWithDuration:current_duration ISO:new_iso completionHandler:Nil]
+                        msg_send![&self.inner, setExposureModeCustomWithDuration:*current_duration ISO:new_iso completionHandler:Nil]
                     };
 
                     Ok(())
@@ -1046,7 +997,7 @@ mod internal {
                     }
 
                     let _: () = unsafe {
-                        msg_send![self.inner, setExposureModeCustomWithDuration:new_duration ISO:current_iso completionHandler:Nil]
+                        msg_send![&self.inner, setExposureModeCustomWithDuration:new_duration ISO:current_iso completionHandler:Nil]
                     };
 
                     Ok(())
@@ -1094,7 +1045,7 @@ mod internal {
                         });
                     }
 
-                    let _: () = unsafe { msg_send![self.inner, whiteBalanceMode: setter] };
+                    let _: () = unsafe { msg_send![&self.inner, whiteBalanceMode: setter] };
 
                     Ok(())
                 }
@@ -1136,7 +1087,7 @@ mod internal {
                         });
                     }
 
-                    let _: () = unsafe { msg_send![self.inner, whiteBalanceMode: setter] };
+                    let _: () = unsafe { msg_send![&self.inner, whiteBalanceMode: setter] };
 
                     Ok(())
                 }
@@ -1179,7 +1130,7 @@ mod internal {
                         });
                     }
 
-                    let _: () = unsafe { msg_send![self.inner, whiteBalanceMode: setter] };
+                    let _: () = unsafe { msg_send![&self.inner, whiteBalanceMode: setter] };
 
                     Ok(())
                 }
@@ -1221,7 +1172,7 @@ mod internal {
                     }
 
                     let _: () = unsafe {
-                        msg_send![self.inner, rampToVideoZoomFactor: setter withRate: 1.0_f32]
+                        msg_send![&self.inner, rampToVideoZoomFactor: setter withRate: 1.0_f32]
                     };
 
                     Ok(())
@@ -1264,7 +1215,7 @@ mod internal {
                         });
                     }
 
-                    let _: () = unsafe { msg_send![self.inner, exposureMode: setter] };
+                    let _: () = unsafe { msg_send![&self.inner, exposureMode: setter] };
 
                     Ok(())
                 }
@@ -1311,7 +1262,7 @@ mod internal {
                         });
                     }
 
-                    let _: () = unsafe { msg_send![self.inner, focusMode: setter] };
+                    let _: () = unsafe { msg_send![&self.inner, focusMode: setter] };
 
                     Ok(())
                 }
@@ -1359,7 +1310,7 @@ mod internal {
                             });
                         }
 
-                        let _: () = unsafe { msg_send![self.inner, focusPointOfInterest: setter] };
+                        let _: () = unsafe { msg_send![&self.inner, focusPointOfInterest: setter] };
 
                         Ok(())
                     }
@@ -1401,7 +1352,7 @@ mod internal {
                         }
 
                         let _: () = unsafe {
-                            msg_send![self.inner, setFocusModeLockedWithLensPosition: setter handler: Nil]
+                            msg_send![&self.inner, setFocusModeLockedWithLensPosition: setter handler: Nil]
                         };
 
                         Ok(())
@@ -1450,7 +1401,7 @@ mod internal {
                         }
 
                         let _: () =
-                            unsafe { msg_send![self.inner, exposurePointOfInterest: setter] };
+                            unsafe { msg_send![&self.inner, exposurePointOfInterest: setter] };
 
                         Ok(())
                     }
@@ -1543,7 +1494,7 @@ mod internal {
                         }
 
                         let _: () = unsafe {
-                            msg_send![self.inner, setExposureTargetBias: setter handler: Nil]
+                            msg_send![&self.inner, setExposureTargetBias: setter handler: Nil]
                         };
 
                         Ok(())
@@ -1587,7 +1538,7 @@ mod internal {
                             });
                         }
 
-                        let _: () = unsafe { msg_send![self.inner, torchMode: setter] };
+                        let _: () = unsafe { msg_send![&self.inner, torchMode: setter] };
 
                         Ok(())
                     }
@@ -1634,7 +1585,7 @@ mod internal {
                         }
 
                         let _: () = unsafe {
-                            msg_send![self.inner, geometricDistortionCorrectionEnabled: setter]
+                            msg_send![&self.inner, geometricDistortionCorrectionEnabled: setter]
                         };
 
                         Ok(())
@@ -1654,9 +1605,11 @@ mod internal {
         }
 
         pub fn active_format(&self) -> Result<CameraFormat, NokhwaError> {
-            let avf_format: AVCaptureDeviceFormat = unsafe { msg_send![self.inner, activeFormat] };
+            let avf_format: AVCaptureDeviceFormat = unsafe { msg_send![&self.inner, activeFormat] };
             let resolution = avf_format.resolution;
-            let fourcc = avf_format.fourcc;
+            let media_type = avf_format.media_type();
+            let media_type_bytes = media_type.to_string().as_bytes();
+            let fourcc = FourCC::from(media_type_bytes);
             let mut a = avf_format
                 .fps_list
                 .into_iter()
@@ -1678,36 +1631,6 @@ mod internal {
                     error: "None??".to_string(),
                 })
             }
-        }
-    }
-
-    pub struct AVCaptureVideoDataOutput {
-        inner: *mut AnyObject,
-    }
-
-    impl AVCaptureVideoDataOutput {
-        pub fn new() -> Self {
-            AVCaptureVideoDataOutput::default()
-        }
-
-        pub fn add_delegate(&self, delegate: &AVCaptureVideoCallback) -> Result<(), NokhwaError> {
-            unsafe {
-                let _: () = msg_send![
-                    self.inner,
-                    setSampleBufferDelegate: delegate.delegate
-                    queue: delegate.queue()
-                ];
-            };
-            Ok(())
-        }
-    }
-
-    impl Default for AVCaptureVideoDataOutput {
-        fn default() -> Self {
-            let cls = class!(AVCaptureVideoDataOutput);
-            let inner: *mut AnyObject = unsafe { msg_send![cls, new] };
-
-            AVCaptureVideoDataOutput { inner }
         }
     }
 }